+++ src/epoll.hpp
@@ -50,7 +50,7 @@ struct i_poll_events;
 //  This class implements socket polling mechanism using the Linux-specific
 //  epoll mechanism.
 
-class epoll_t : public poller_base_t
+class epoll_t : public worker_poller_base_t
 {
   public:
     typedef void *handle_t;
@@ -65,20 +65,13 @@ class epoll_t : public poller_base_t
     void reset_pollin (handle_t handle_);
     void set_pollout (handle_t handle_);
     void reset_pollout (handle_t handle_);
-    void start ();
     void stop ();
 
     static int max_fds ();
 
   private:
-    //  Main worker thread routine.
-    static void worker_routine (void *arg_);
-
     //  Main event loop.
     void loop ();
-
-    // Reference to ZMQ context.
-    const thread_ctx_t &ctx;
 
     //  Main epoll file descriptor
     fd_t epoll_fd;
@@ -93,9 +86,6 @@ class epoll_t : public poller_base_t
     //  List of retired event sources.
     typedef std::vector<poll_entry_t *> retired_t;
     retired_t retired;
-
-    //  If true, thread is in the process of shutting down.
-    bool stopping;
 
     //  Handle of the physical thread doing the I/O work.
     thread_t worker;
